refactor: Split ReservationUnit edit page components to separate files
